+++ database/migrations/2023_03_02_051405_create_transaksis_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('kode_transaksi', 10)->unique();
             $table->date('tgl_transaksi');
-            $table->bigInteger('total_transaksi');
+            $table->decimal('total_transaksi', 30, 3);
             $table->timestamps();
         });
     }
